Give the Python-object value contract its own header and impl file
Review finding on #663: the four primitives were declared in
bridge_state.h and implemented in bridge_state.cpp, mixing a reusable,
stateless value contract with the registry and slot state that file
owns, so a semantic owner (the retained-value entry in types/metadata,
the bridge's PyObj hash) depended on the bridge-state facade to reach
them (AGENTS.md, "Type-Erased Strategy Boundaries").

include/hgraph/python/object_semantics.h now carries the contract and
its rules; src/hgraph/python/impl/object_semantics.cpp the CPython
implementation. The Python-owned Bundle entry renders its diagnostic
type name through object_str, which retires bridge_state.cpp's private
UTF-8 helper. The ratchet owner names the header and the impl file.

Design record: docs/source/developer_guide/python_bridge.rst (module
table row; "One set of Python-object value primitives").
