+++ include/hgraph/python/bridge_state.h
@@ -55,23 +55,6 @@ struct HGRAPH_LOCAL PythonValueHolder {
   }
 };
 
-/**
- * The one set of Python-object value primitives shared by every ops table
- * that stores a PyObject (Python-owned Bundles, retained values, the
- * bridge's PyObj hash). Hash falls back to the object's address when the
- * object is unhashable; equality runs Python ``__eq__`` first so its
- * exceptions stay observable, then refuses to call two unhashable objects
- * equal so equal values keep equal hashes; ordering maps a Python comparison
- * error to ``unordered`` and an empty side below a live one. Each acquires
- * the GIL itself. ``object_hash`` requires a live object; callers decide
- * what an empty holder means.
- */
-[[nodiscard]] HGRAPH_EXPORT std::size_t object_hash(PyObject *object);
-[[nodiscard]] HGRAPH_EXPORT bool object_equals(PyObject *lhs, PyObject *rhs);
-[[nodiscard]] HGRAPH_EXPORT std::partial_ordering object_compare(PyObject *lhs,
-                                                                PyObject *rhs) noexcept;
-[[nodiscard]] HGRAPH_EXPORT std::string object_str(PyObject *object);
-
 [[nodiscard]] HGRAPH_EXPORT nb::object &cmp_result_enum_slot();
 [[nodiscard]] HGRAPH_EXPORT nb::object &divide_by_zero_enum_slot();
 /** The recording-shape policy enums (RFC 0019). Registered like the enums
