+++ src/CMakeLists.txt
@@ -96,6 +96,7 @@ set(HGRAPH_RUNTIME_SOURCES
 if(HGRAPH_BUILD_PYTHON_BINDINGS OR HGRAPH_ENABLE_PYTHON_USER_NODES)
     list(APPEND HGRAPH_RUNTIME_SOURCES
         hgraph/python/bridge_state.cpp
+        hgraph/python/impl/object_semantics.cpp
         hgraph/python/impl/ts_data_conversion.cpp
     )
 endif()
